moved all compat code to rdflib.compat
